Fix compatibility issues with MPS on Apple silicon
- TypeError: Cannot convert a MPS Tensor to float64 dtype as the MPS framework doesn't support float64. Please use float32 instead.
- RuntimeError: MPS does not support min/max ops with int64 input
- NotImplementedError: The operator 'torchvision::nms' is not currently implemented for the MPS device. If you want this op to be added in priority during the prototype phase of this feature, please comment on https://github.com/pytorch/pytorch/issues/77764. As a temporary fix, you can set the environment variable `PYTORCH_ENABLE_MPS_FALLBACK=1` to use the CPU as a fallback for this op. WARNING: this will be slower than running natively on MPS.
- RuntimeError: dets should have the same type as scores

+++ mobile_sam/automatic_mask_generator.py
@@ -277,7 +277,7 @@ class SamAutomaticMaskGenerator:
                 raise NotImplementedError()
             keep_by_nms = batched_nms(
                 data["boxes"].float(),
-                scores.to(data["boxes"].device),
+                scores.to(data["boxes"].device).float(),
                 torch.zeros_like(data["boxes"][:, 0]),  # categories
                 iou_threshold=self.box_nms_thresh,
             )
